cart remove and add affect cart list, checks do not persist in database though

+++ models/tables.py
@@ -54,7 +54,7 @@ db.define_table('images',
                 Field('created_on', 'datetime', default=request.now),
                 Field('created_by', 'reference auth_user', default=auth.user_id),
                 Field('image_url'),
-                Field('image_price','integer'),
+                Field('image_price','float'),
                 Field('is_checked', 'boolean', default=False),
 )
